Create resutls module and move QueryResult here
Create separate module for results of DetaBase methods
