+++ deta_py/deta_base/async_base.py
@@ -8,7 +8,8 @@ from typing import Any, Optional
 
 from aiohttp import ClientSession, ClientTimeout
 
-from deta_py.deta_base.queries import ItemUpdate, QueryResult
+from deta_py.deta_base.queries import ItemUpdate
+from deta_py.deta_base.results import QueryResult
 from deta_py.deta_base.types import ExpireAt, ExpireIn, Query
 from deta_py.deta_base.utils import (
     BASE_API_URL,
