+++ deta_py/deta_base/base.py
@@ -11,7 +11,8 @@ from typing import Any, Optional
 
 from requests import Response, Session
 
-from deta_py.deta_base.queries import ItemUpdate, QueryResult
+from deta_py.deta_base.queries import ItemUpdate
+from deta_py.deta_base.results import QueryResult
 from deta_py.deta_base.types import ExpireAt, ExpireIn, Query
 from deta_py.deta_base.utils import (
     BASE_API_URL,
